// hivemq-edge-frontend/src/api/__generated__/models/ManagedAssetList.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ManagedAsset } from './ManagedAsset';

/**
 * A list of managed assets, as managed in Edge for the Pulse Client
 */
export type ManagedAssetList = {
    items: Array<ManagedAsset>;
};